test: reproduce concurrent timeout repair overwrite
